--- a/Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
@@ -398,12 +398,13 @@ namespace Jint.Runtime.Interpreter.Expressions
                 var engine = context.Engine;
                 var env = engine.ExecutionContext.LexicalEnvironment;
                 var strict = StrictModeScope.IsStrictModeCode;
+                var identifier = left.Identifier;
                 if (JintEnvironment.TryGetIdentifierEnvironmentWithBinding(
                     env,
-                    left.Identifier,
+                    identifier,
                     out var environmentRecord))
                 {
-                    if (strict && hasEvalOrArguments && left.Identifier.Key != KnownKeys.Eval)
+                    if (strict && hasEvalOrArguments && identifier.Key != KnownKeys.Eval)
                     {
                         ExceptionHelper.ThrowSyntaxError(engine.Realm, "Invalid assignment target");
                     }
@@ -418,10 +419,10 @@ namespace Jint.Runtime.Interpreter.Expressions
 
                     if (right._expression.IsFunctionDefinition())
                     {
-                        ((FunctionInstance) rval).SetFunctionName(left.Identifier.Value);
+                        ((FunctionInstance) rval).SetFunctionName(identifier.Value);
                     }
 
-                    environmentRecord.SetMutableBinding(left.Identifier, rval, strict);
+                    environmentRecord.SetMutableBinding(identifier, rval, strict);
                     return rval;
                 }