ArrayStressBenchmark:
``` ini

BenchmarkDotNet=v0.13.5, OS=Windows 10 (10.0.19045.3086/22H2/2022Update)
Intel Core i7-4720HQ CPU 2.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.302
  [Host]     : .NET 6.0.16 (6.0.1623.17311), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.16 (6.0.1623.17311), X64 RyuJIT AVX2
```
main:
|               Method |     Mean |    Error |   StdDev |      Gen0 |     Gen1 | Allocated |
|--------------------- |---------:|---------:|---------:|----------:|---------:|----------:|
|              Execute | 24.45 ms | 0.132 ms | 0.123 ms | 2250.0000 | 187.5000 |   7.18 MB |
| Execute_ParsedScript | 22.74 ms | 0.106 ms | 0.094 ms | 2312.5000 |  93.7500 |   7.15 MB |


PR:
|               Method |     Mean |    Error |   StdDev |      Gen0 |     Gen1 | Allocated |
|--------------------- |---------:|---------:|---------:|----------:|---------:|----------:|
|              Execute | 22.85 ms | 0.106 ms | 0.088 ms | 2250.0000 | 187.5000 |   7.18 MB |
| Execute_ParsedScript | 21.95 ms | 0.076 ms | 0.064 ms | 2250.0000 | 250.0000 |   7.15 MB |